requirements.txt
mcp[cli]>=0.7.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
pip-system-certs>=0.1.0 